+++ Runtime/Settings/IGameManagerSettingsProfile.cs
@@ -1,8 +1,8 @@
 ﻿using UnityEngine;
 
-namespace CHARK.GameManagement
+namespace CHARK.GameManagement.Settings
 {
-    public interface IGameManagerSettings
+    internal interface IGameManagerSettingsProfile
     {
         /// <summary>
         /// Should <see cref="GameManager"/> prefab instantiate automatically?
@@ -13,7 +13,7 @@ namespace CHARK.GameManagement
         /// When to load <see cref="GameManager"/> prefab when instantiating it automatically. Used
         /// only when <see cref="IsInstantiateAutomatically"/> is set to <c>true</c>.
         /// </summary>
-        public RuntimeInitializeLoadType LoadType { get; }
+        public InstantiationMode InstantiationMode { get; }
 
         /// <summary>
         /// Should <see cref="GameManager"/> use <see cref="Object.DontDestroyOnLoad"/>.